双读部分代码提交： 1.登录部分校验用户名和密码双读代码 2.自测用例从double_read分支移动到both_read分支 3.部分代码缓存改为DB缓存

+++ passport-main-common/src/main/java/com/sogou/upd/passport/common/utils/DBShardRedisUtils.java
@@ -79,15 +79,16 @@ public class DBShardRedisUtils {
     }
 
 
-    public String set(final String key, final String value, final int expire) {
+    public String set(final String key, final String value, final long expire) {
         return new Executor<String>(shardedJedisPool) {
 
             @Override
             String execute() {
-                return jedis.setex(key, expire, value);
+                return jedis.setex(key, (int) expire, value);
             }
         }.getResult();
     }
+
 
     @Profiled(el = true, logger = "rediesTimingLogger", tag = "dbShardRedis_setWithinSeconds", timeThreshold = 10, normalAndSlowSuffixesEnabled = true)
     public String setWithinSeconds(final String key, final Object obj, final long timeout) {
